requirements.txt
tweepy
google-genai
requests
python-dotenv
feedparser
matplotlib
pandas
seaborn
